chore: extract broadcast suite into src/suites/broadcast.ts
Part of REAL-1037.
