+++ test/e2e/realtime-check.ts
@@ -23,6 +23,7 @@ import { loadBroadcastFromDb } from "./src/suites/load-broadcast-from-db.ts";
 import { loadBroadcastReplay } from "./src/suites/load-broadcast-replay.ts";
 import { connection } from "./src/suites/connection.ts";
 import { loadPresence } from "./src/suites/load-presence.ts";
+import { broadcast } from "./src/suites/broadcast.ts";
 
 async function runLoadPostgresChangesTests(testUser: { email: string; password: string }) {
   suite("load-postgres-changes");
@@ -209,64 +210,6 @@ async function runLoadBroadcastTests() {
   });
 }
 
-
-async function runBroadcastTests() {
-  suite("broadcast extension");
-
-  await test("user is able to receive self broadcast", async () => {
-    const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    try {
-      let result: any = null;
-      const event = crypto.randomUUID();
-      const topic = randomTopic();
-      const expectedPayload = { message: crypto.randomUUID() };
-
-      const channel = supabase
-        .channel(topic, BROADCAST_CONFIG)
-        .on("broadcast", { event }, ({ payload }) => (result = payload));
-
-      const subscribeMs = await openChannel(channel);
-      await channel.send({ type: "broadcast", event, payload: expectedPayload });
-      const { latencyMs: eventMs } = await waitFor(() => result, "broadcast event");
-
-      assert.deepStrictEqual(result, expectedPayload);
-      return [{ label: "subscribe", value: subscribeMs, unit: "ms" }, { label: "event", value: eventMs, unit: "ms" }];
-    } finally {
-      await stopClient(supabase);
-    }
-  });
-
-  await test("user is able to use the endpoint to broadcast", async () => {
-    const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    try {
-      let result: any = null;
-      const event = crypto.randomUUID();
-      const topic = randomTopic();
-      const expectedPayload = { message: crypto.randomUUID() };
-
-      const channel = supabase
-        .channel(topic, BROADCAST_CONFIG)
-        .on("broadcast", { event }, ({ payload }) => (result = payload));
-
-      const subscribeMs = await openChannel(channel);
-      // Small settle window so server-side subscription routing is ready before the HTTP broadcast arrives.
-      await sleep(100);
-
-      const res = await fetch(`${PROJECT_URL}/realtime/v1/api/broadcast`, {
-        method: "POST",
-        headers: BROADCAST_API_HEADERS,
-        body: JSON.stringify({ messages: [{ topic, event, payload: expectedPayload }] }),
-      });
-      if (!res.ok) throw new Error(`Broadcast API returned ${res.status}`);
-
-      const { latencyMs: eventMs } = await waitFor(() => result, "broadcast event");
-      assert.deepStrictEqual(result, expectedPayload);
-      return [{ label: "subscribe", value: subscribeMs, unit: "ms" }, { label: "event", value: eventMs, unit: "ms" }];
-    } finally {
-      await stopClient(supabase);
-    }
-  });
-}
 
 async function runPresenceTests(_testUser: { email: string; password: string }, supabase: SupabaseClient) {
   suite("presence extension");
@@ -1249,7 +1192,7 @@ const descriptors: SuiteDescriptor[] = [
   { name: "load-broadcast", label: "load-broadcast", needsDb: false, run: () => runLoadBroadcastTests() },
   loadBroadcastFromDb,
   loadBroadcastReplay,
-  { name: "broadcast", label: "broadcast extension", needsDb: false, run: () => runBroadcastTests() },
+  broadcast,
   { name: "broadcast-replay", label: "broadcast replay", needsDb: true, run: ({ testUser, supabase }) => runBroadcastReplayTests(testUser, supabase) },
   { name: "presence", label: "presence extension", needsDb: true, run: ({ testUser, supabase }) => runPresenceTests(testUser, supabase) },
   authorization,
